backend/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pandas
db-dtypes
